🐛 Fix "React Native platform is undefined"
Due to a strange bug in Expo 50, I had to move the import of the Platform lib from the top of the file to the bottom.
I don't know why it works, but let's keep it that way.

By looking at imports of Platform, I've noticed duplicate code so I deleted it (menu bar color was updated in app.js and in the home screen)